move App.jsx to components folder
