+++ styles/App.css
@@ -1,0 +1,4 @@
+body {
+  background-color: pink;
+}
+/*# sourceMappingURL=App.css.map */
